réinsertion de createUser, generateInvitation et checkUser dans le fichier Ardennes
