+++ .env.production
@@ -1,7 +1,7 @@
 NEXT_PUBLIC_MATOMO_URL=https://stats.data.gouv.fr
 NEXT_PUBLIC_MATOMO_SITE_ID=158
-NEXT_PUBLIC_RDV_SOLIDARITES_DEMO_URL=https://demo.rdv-solidarites.fr/
-NEXT_PUBLIC_RDV_SOLIDARITES_PROD_URL=https://rdv-solidarites.fr/
+NEXT_PUBLIC_RDV_SOLIDARITES_DEMO_URL=https://demo.rdv-solidarites.fr
+NEXT_PUBLIC_RDV_SOLIDARITES_PROD_URL=https://rdv-solidarites.fr
 NEXT_PUBLIC_ORGANISATION_ID_DEMO=164
 NEXT_PUBLIC_ORGANISATION_ID_PROD=249
