+++ lib/appFetch.js
@@ -9,7 +9,7 @@ export async function appFetch(url, token, method = "GET", body = null) {
     },
     body: body,
   })
-  const result = response.json();
+  const result = await response.json();
 
   return result;
 };
